// front-06-test/src/components/units/board/show/boardShow.types.ts
import { IQuery } from "../../../../commons/types/generated/types";

//UI
export interface IBoardRoadUIProps {
  data?: any;
  OnClickList: () => void;
  OnClickEdit: () => void;
  OnClickDelete: () => void;

  isModal: boolean;
  showModal: () => void;
}
//컨테이너
export interface INewBoardConProps {
  isEdit: boolean;
  //     data : Pick<IMutation, "createBoard">
  data: Pick<IQuery, "fetchBoard">;
}